+++ generator/generator.js
@@ -88,6 +88,7 @@ require("./page_generators/documentation/panther/literals.js");
 require("./page_generators/documentation/panther/modules.js");
 require("./page_generators/documentation/panther/operators.js");
 require("./page_generators/documentation/panther/primitive_types.js");
+require("./page_generators/documentation/panther/structs.js");
 require("./page_generators/documentation/panther/undefined_behavior.js");
 require("./page_generators/documentation/panther/uninitialized.js");
 require("./page_generators/documentation/panther/value_categories.js");